Refactor project structure and enhance configuration management
Relocated `helpers` into `cmd` for improved organization and streamlined imports. Updated configuration handling with clearer examples and logging adjustments. Added `make install` and `make uninstall` targets for easier setup and cleanup.
