+++ Makefile
@@ -5,6 +5,7 @@ GOTEST     = $(GOCMD) test
 GOFMT      = $(GOCMD) fmt
 GOVET      = $(GOCMD) vet
 GOMOD      = $(GOCMD) mod
+GOINSTALL  = $(GOCMD) install
 
 # The name of the output binary. Adjust if your main package uses a different name.
 BINARY_NAME = cloudflare-dyndns
@@ -18,13 +19,19 @@ TARGETS ?= darwin linux windows
 # Set the build architecture.
 ARCH ?= amd64 arm64
 
-.PHONY: all release build test run fmt vet clean tidy verify
+.PHONY: all release build test run fmt vet clean tidy verify install release-all
 
-# Default target builds the binary.
-all: build
+install:
+	@echo "Installing $(BINARY_NAME) for your system"
+	GOOS=$(shell go env GOOS) GOARCH=$(shell go env GOARCH) $(GOINSTALL) -trimpath -ldflags="-s -w" .
+	@echo "Done! Don't forget to create your configuration file (see README.md) before running."
+
+uninstall:
+	rm "$(shell go env GOPATH)/bin/$(BINARY_NAME)"
+	@echo "Done! Don't forget to delete any configuration file you created."
 
 # Release compiles an optimized version of the binary.
-release: vet verify test clean
+release-all: vet verify test clean
 	@echo "Building binaries for: $(TARGETS)"
 	@for os in $(TARGETS); do \
 		for arch in $(ARCH); do \
@@ -37,6 +44,15 @@ release: vet verify test clean
 			GOOS=$$os GOARCH=$$arch $(GOBUILD) -trimpath -ldflags="-s -w" -o $(BUILD_DIR)/$$os/$$arch/$(BINARY_NAME)$$ext .; \
 		done; \
 	done
+
+release: vet verify test clean
+	@os="$(shell go env GOOS)"
+	@arch="$(shell go env GOARCH)"
+	@ext="";
+	@if [ "$$os" = "windows" ]; then \
+		ext=".exe"; \
+	fi;
+	GOOS=$$os GOARCH=$$arch $(GOBUILD) -trimpath -ldflags="-s -w" -o $(BUILD_DIR)/$$os/$$arch/$(BINARY_NAME)$$ext .;
 
 # Build compiles the Go code and outputs the binary into the build directory.
 build:
