+++ cmd/helpers_test.go
@@ -1,4 +1,4 @@
-package helpers
+package cmd
 
 import (
 	"bytes"
