+++ constants/constants.go
@@ -1,6 +1,5 @@
 package constants
 
 const (
-	Version  = "0.0.1"
 	MaxTries = 3
 )
